Remove functionality related to the link toggle option
Remove functionality related to the link toggle option.

Do not show the link option in the toolbar when the featured image is used.

Reset the link option when an image that is linked is replaced with the featured image option.
Without this reset, there is a link output on the front, in an incorrect position in the markup.