Block Variations: Detect active variation correctly based on RichText attribute (#62325)
If a block variation has an `isActive` property that is an array of strings (i.e. attribute names), and one of the attributes referenced therein is a `RichText` value, compare the given `RichText` string correctly to the given block's. (This applies also to nested attributes, i.e. if a `RichText` value is contained in an object (that might be contained in another, and so on) inside an attribute.)

Co-authored-by: ockham <bernhard-reiter@git.wordpress.org>
Co-authored-by: cbravobernal <cbravobernal@git.wordpress.org>
Co-authored-by: gziolo <gziolo@git.wordpress.org>