i18n: Add context to "Sync this pattern" label
N.B.: I may revert this commit in favour of simply fixing the string in
my locale, but wanted to share the problem first.

In my testing locale (pt-PT) I found this label unexpectedly rendered in
the imperative mood ("Sincronize este padrão […]"). Since this string
describes a user action mediated by a checkbox, I would expect the
infinitive instead ("Sincronizar este padrão […]").

That said, we have a lot of strings like this throughout the editor(s),
and lack of context is generally not a problem: the strings speak for
themselves.