Cherry pick script: make authenticated requests
During the 6.6 Beta 2 release, we ran into rate limiting issues due to
conflicts that had to be resolved, which meant that we had to run the
script multiple times.

Non-authenticated requests have a 60 per hour limit.

In order to avoid it in the future, we can grab the token generated by `gh` and use
it to make authenticated requests instead, which have a 5000-per hour
limit.

```sh
➜  gutenberg git:(trunk) curl -vvvv --request GET \
--url "https://api.github.com/octocat" \
--header "Authorization: Bearer <TOKEN_GENERATED_BY_GH_CLI>" \
--header "X-GitHub-Api-Version: 2022-11-28"
...
< HTTP/2 200
< server: GitHub.com
< date: Tue, 11 Jun 2024 10:36:15 GMT
< content-type: application/octocat-stream
< content-length: 917
< x-github-api-version-selected: 2022-11-28
< x-ratelimit-limit: 5000
< x-ratelimit-remaining: 4992
```