Enable skipDefaultLibCheck
This skips _checking_ types that are bundled with TypeScript. We should
be able to trust these types, so they can be used without checking them.

Many packages use the DOM types and this appeared as a hot spot when
analyzing the TypeScript build.

Adding this configuration should improve TypeScript build performance.