Compat: Update HTML API with changes from 6.6 (#63089)
Updates the HTML API components in the compatability layer with updates
from each of the 6.4, 6.5, and 6.6 branches from Core.

Notably, because of the breaking change in 6.5 to the `WP_HTML_Span` and
`WP_HTML_Text_Replacement` classes, which changed the range format from
`(start, end)` to `(start, length)`, the 6.6 classes still reference the
`Gutenberg_HTML_Span_6_5` and `Gutenberg_HTML_Text_Replacement_6_5`
substitutions.

While updating it was noticed that the 6.4 compatability code
accidentally pulled in updates from the 6.5 branch for the HTML
Processor. These additions have been removed for the sake of closer
correspondance with the Core code.

Co-authored-by: ockham <bernhard-reiter@git.wordpress.org>