Separate Command Palette (command + K) and Visible Template title buttons
This is an odd one to keep the same design and account for improved accessibility. It's admittedly a hodgepodge. The shortcut button (Command + K) is a full-width button with the text aligned right. It has an aria-label tooltip of Command Palette so it can be accessed via a voice command of Command Palette or via the visible template name. The template name button is tabindex=-1 so it never receives focus and is also _behind_ the transparent Command + K button, so that a mouse click will always access the Command + K button and get the full width focus style. The H1 for the template name has been changed to a span. I'll add a visually hidden H1 in a follow-up commit.