Scripts: Include variations paths in build (#63098)
If a `block.json`'s `variations` field is set to a (PHP) file name, include that file in the build, i.e. prefix its functions with `gutenberg_` and copy the resulting file to the `build/block-library/blocks/` directory.

Note that this applies to third-party blocks, but not Core blocks. Making it work for Core blocks is a separate issue: #63077.

Co-authored-by: ockham <bernhard-reiter@git.wordpress.org>
Co-authored-by: SantosGuillamot <santosguillamot@git.wordpress.org>