Stablize the border supports
This commit references the PR - https://github.com/WordPress/wordpress-develop/pull/7069/files, which aims to stablise the typogrpahy supports. Based on the PR added this commit to stablise the supports for the border. Since we does not have the border options as experimental, instead we have border as experimiental hence we stablised the main key and this makes us not to touch the inner border supports as they were not defined as experimental.