Remove invisible overlay over metaboxes while saving
An invisible element is placed on top of metaboxes in the block editor, preventing clicks while the post is saving. While this might make sense in theory, in reality, it results in flaky playwright tests which rely on access to elements in that element. 

It's also inconsistently used, as blocks themselves do not do this same thing. 

Removing this makes playwright tests less flaky, provides a more consistent user experience.

Applying invisible divs preventing clicks are not a reliable way to doing this anyway, as you can still use the tab key to navigate into that area and make changes, for example.