Remove unnecessary ToolbarGroup wrappers
There are several toolbar items that were a singular button. I'm wondering if this was a mistake, as every button was wrapped in a group. I'm removing all the groups with only one item to see if everything works the same in order to improve semantics and reduce some unnecessary HTML.