Add editorMode to blockProps context to reduce useSelect and event listeners
Reducing the number of useSelect calls and event listeners is important for performance. Rather than have a useSelect to check zoom out mode and add event listeners to every block, let's do the useSelect once in the provider and pass it in to the hook, and only add the event listener if we're in zoom out mode.