Don't set contentEditableWrapper as true unless multiselect
The contentEditableWrapper is getting toggled to true when still on a single block selection, which moves focus to the wrapper and causes screen readers to announce focus in the wrong location.

This changes to only set that content editable wrapper when multiselecting so selection within the same block doesn't move focus.