Update to use a11y script module package in Core. (#65539)
Update the interactivity-router package to use the @wordpress/a11y module for its aria-live regions.

This was behind a Gutenberg-only flag. Since the a11y module is landing in Core in 6.7, the flag can be removed to always use the a11y module.

See https://github.com/WordPress/wordpress-develop/pull/7304.