Iframe takes up entire zoom out gray background area
The iframe was previously offset by a margin left so that it did not take up the entirety of the gray area. This caused a a couple issues:
- zoom out inserter and toolbars were misaligned when the area was reized
- the scroll bar for the canvas was mysteriously floating in the gray area

By making the iframe take up the full width, it makes the scrollbar set to the right side of the frame and also does not reposition the zoom out inserters or toolbar so they stay in the right place.