Unwrap custom dataviews empty element from paragraph tags
The custom empty element can be anything arbitrary, including divs and
headings. These elements are not semantically allowed to be children of
paragraphs.

The spinner continues to be wrapped in a paragraph. This is allowed
semantically (the spinner is just an <svg> element), it keeps the
vertical margins of the spinner consistent with the default empty
message, and it is consistent with the spinners which show when the
dataviews are loading more data in infinite scroll mode.