Image block: Conditionally remove empty `<figcaption>` (#71893)
Remove the `<figcaption>` element from the block markup upon render if the `caption` attribute is empty.

Since `caption` is a sourced attribute, this requires checking if the `<figcaption>` has any inner elements. If it doesn't, it's still possible that the `caption` attribute is connected to a block bindings source, so we also need to rule out that it received a non-empty value from that source. If neither criterion is met, we remove the `<figcaption>`.

Co-authored-by: ockham <bernhard-reiter@git.wordpress.org>
Co-authored-by: justintadlock <greenshady@git.wordpress.org>
Co-authored-by: dmsnell <dmsnell@git.wordpress.org>
Co-authored-by: gziolo <gziolo@git.wordpress.org>