Build: Support watching SVG icons to generate TSX
Follow-up of #71878

As of #71878, the library of icons in the Icons package consists of
plain SVG files, from which a build step generates TSX files and an
index.ts.

However, this generation only happened once upon invoking `npm run dev`
(and `npm run build`). This commit adds support for watching changes to
those SVG files and regenerating accordingly.