Don't apply selection change for off-canvas text length changes
The attempt to keep track of caret position (applyRange) was stealing focus back to the canvas when those changes were coming from an off-canvas input (navigation link block text changes from inspector sidebar input). In these cases, it's not an expected behavior to keep the selection location, as the text length is changing intentionally off canvas. In these cases, it's better to not keep track of that focus.