Icons: Bundle generated source code
As of #71878, Icons are defined as SVG files under `src/library/`, and
an accompanying `build` script will produce their corresponding TSX
files and an index at `src/library/index.ts`.

However, this generated source code was not included in the published
package. This pull request seeks to include `src/library/*.ts[x]` in
`@wordpress/icons`, for the benefit of its consumers, who thereby no
longer need to generate that code themselves.

It does so by adding an empty `.npmignore` file, which overrides
`.gitignore`. As a reminder, we deliberately Git-ignore the generated
source code because we don't want the duplication of SVG and TSX files,
nor the possible synchronisation issues that come with it.

@see https://docs.npmjs.com/cli/v11/commands/npm-publish#files-included-in-package