Fix empty url value from unbinding entity from inspector sidebar
Unbinding the navigation link entity from the sidebar updates the url value to be '', which causes the canvas to go into its empty URL state, which removes the label. Then, when you start typing in the url link input, it would steal focus back to the canvas. This changes the flow when unbinding to set the url as the previously linked url, which avoids an empty link issue. However, you can still get to the messed up state by deleting the url value. To fix this, instead of updating the attribute onChange, update it onBlur so we can restore the last known url.