Remove useBlockDisplayTitle from BlockSwitcher
useBlockDisplayTitle created two subscriptions and was running everytime content length changed. The main reason to use that hook is if you need to display the custom block title, which we don't. So, by switching to match?.title and blockType.title, we remove two subscriptions that were firing many times each keystroke in order to get a label for a button that isn't changing.