Ignore generated icon files for ESLint
These files are auto-generated, and expected to have some formatting issues. However, blanket disabling Prettier is incompatible with no-unused-disable because some of the files are fine as-is.

A better long-term solution would be to apply Prettier to the generated output, similar to what's done in `packages/theme` for Theme artifacts.