Alt: remove bound navigation links from editor if their synced source is deleted
Rather than display UI states to resolve broken links, the user likely wants the link entirely removed. So, instead of displaying a state they need to resolve, let's remove the links for them. If a link is supposed to be bound, we check if the source exists. While a synced navigation link is resolving, we hide it from display. Once it is resolved and valid, we display it.