Fix a11y of descriptions and alerts for "Invalid" Nav Items (#73177)
The role="alert" and aria-live="polite" attributes were causing the error message to be announced twice to screen readers. Since this text is referenced via aria-describedby on the input field, it will be read when users focus the field - no live announcement is needed.

This also adds aria-invalid and aria-describedBy to the navigation link block if the synced entity is not found. This allows screen reader users to be aware of the issue on canvas as well.

---------

Co-authored-by: Dave Smith <getdavemail@gmail.com>