Block Bindings: Move computation of sources into dedicated `useSelect` (#72974)
There are a few hacks and workarounds in Block Bindings to make them work. One example is the definition of `const _sources` outside of the following `useSelect` statement, which was done to avoid a console warning ("The `useSelect` hook returns different values when called with the same state and parameters").

Additionally, e2e tests had workarounds to make them pass in CI.

This commit moves `sources` computation into a dedicated `useSelect`, applies some general refactoring, and removes e2e workarounds.

Co-authored-by: ockham <bernhard-reiter@git.wordpress.org>
Co-authored-by: gziolo <gziolo@git.wordpress.org>
Co-authored-by: sirreal <jonsurrell@git.wordpress.org>