Fix block toolbar icon CSS when using show icon label preference
The CSS for the block toolbar icon was refactored to remove it from the block switcher. This refactor caused a minor regression in the styles for the show icon label preference. The CSS should have been scoped to the block toolbar, but was instead scoped to any instance of the block icon. This fix applies the correct scope.