Fix focus management after creating new page in navigation block link ui
The previous LinkControl fix for this was too heavy handed and caused issues with always moving focus into the preview when mounting the link control. I've reverted that and used a more targeted bug fix approach by setting a flag that we need to send focus into the link control after creating a new page.